+++ packages/runtime/src/elements/terminal.ts
@@ -4,48 +4,9 @@ import { Terminal } from "xterm";
 import { WebLinksAddon } from "xterm-addon-web-links";
 import { FitAddon } from "xterm-addon-fit";
 import { WASIFS, WASIWorkerHost, WASIWorkerHostKilledError } from "@runno/wasi";
-import { makeRunnoError } from "./helpers";
+import { makeRunnoError } from "../helpers";
 
-const ATTRIBUTE_MAP = {
-  src: "src",
-  name: "name",
-  args: "args",
-  "disable-echo": "disableEcho",
-  "disable-tty": "disableTTY",
-  controls: "controls",
-  autorun: "autorun",
-} as const;
-
-const BOOLEAN_ATTRIBUTES = [
-  "disable-echo",
-  "disable-tty",
-  "controls",
-  "autorun",
-] as const;
-
-type BooleanAttribute = "disable-echo" | "disable-tty" | "controls" | "autorun";
-
-function isBooleanAttribute(key: string): key is BooleanAttribute {
-  return key in BOOLEAN_ATTRIBUTES;
-}
-
-export class WASIElement extends HTMLElement {
-  static get observedAttributes() {
-    return Object.keys(ATTRIBUTE_MAP);
-  }
-
-  src: string = "";
-  name: string = "program";
-  args: string[] = [];
-  env: Record<string, string> = {};
-  fs: WASIFS = {};
-
-  // Boolean controls
-  disableEcho: boolean = false;
-  disableTTY: boolean = false;
-  controls: boolean = false;
-  autorun: boolean = false;
-
+export class TerminalElement extends HTMLElement {
   // Terminal Display
   terminal: Terminal = new Terminal({
     convertEol: true,
@@ -54,12 +15,13 @@ export class WASIElement extends HTMLElement {
   fitAddon: FitAddon = new FitAddon();
   resizeObserver: ResizeObserver;
 
+  // Configuration Options
+  echoStdin: boolean = true;
+
   // Runtime State
   workerHost?: WASIWorkerHost;
   stdinHistory: string = "";
   ttyHistory: string = "";
-
-  private hasRun = false;
 
   constructor() {
     super();
@@ -73,9 +35,7 @@ export class WASIElement extends HTMLElement {
     this.shadowRoot!.innerHTML = `
     <style>
       :host {
-        display: block;
         position: relative;
-        min-height: 140px;
       }
 
       * {
@@ -98,9 +58,6 @@ export class WASIElement extends HTMLElement {
         left: 0;
         right: 0;
         padding: 0.5em;
-        background: black;
-        height: var(--runno-terminal-height, auto);
-        min-height: var(--runno-terminal-min-height, 4rem);
       }
     </style>
     <div id="container"></div>
@@ -111,9 +68,13 @@ export class WASIElement extends HTMLElement {
   // Public Helpers
   //
 
-  async run(): Promise<RunResult> {
-    this.hasRun = true;
-
+  async run(
+    binaryPath: string,
+    binaryName: string,
+    fs: WASIFS,
+    args: string[],
+    env: { [key: string]: string }
+  ): Promise<RunResult> {
     if (this.workerHost) {
       this.workerHost.kill();
     }
@@ -125,10 +86,10 @@ export class WASIElement extends HTMLElement {
       let stdout = "";
       let stderr = "";
 
-      this.workerHost = new WASIWorkerHost(this.src, {
-        args: [this.name, ...this.args],
-        env: this.env,
-        fs: this.fs,
+      this.workerHost = new WASIWorkerHost(binaryPath, {
+        args: [binaryName, ...args],
+        env,
+        fs,
         isTTY: true,
         stdout: (out) => {
           stdout += out;
@@ -185,27 +146,6 @@ export class WASIElement extends HTMLElement {
     this.resizeObserver.unobserve(this);
   }
 
-  attributeChangedCallback(
-    name: keyof typeof ATTRIBUTE_MAP,
-    _: string,
-    newValue: string
-  ) {
-    if (name === "args") {
-      this.args = newValue.trim() ? newValue.trim().split(" ") : [];
-      return;
-    }
-
-    if (name === "autorun" && !this.hasRun) {
-      this.run();
-    }
-
-    if (isBooleanAttribute(name)) {
-      this[ATTRIBUTE_MAP[name]] = true;
-    } else {
-      this[ATTRIBUTE_MAP[name]] = newValue;
-    }
-  }
-
   //
   // Events
   //
@@ -219,7 +159,7 @@ export class WASIElement extends HTMLElement {
       data = "\n";
     }
 
-    if (!this.disableEcho) {
+    if (this.echoStdin) {
       // TODO: Parse backspace etc
       this.terminal.write(data);
     }
@@ -263,4 +203,4 @@ export class WASIElement extends HTMLElement {
   }
 }
 
-customElements.define("runno-wasi", WASIElement);
+customElements.define("runno-terminal", TerminalElement);
