+++ packages/runtime/src/main.ts
@@ -1,7 +1,8 @@
-export { TerminalElement } from "./terminal";
-export { EditorElement } from "./editor";
-export { ControlsElement } from "./controls";
-export { RunElement } from "./run";
-export { WASIElement } from "./wasi";
+export { TerminalElement } from "./elements/terminal";
+export { EditorElement } from "./elements/editor";
+export { ControlsElement } from "./elements/controls";
+export { RunElement } from "./elements/run";
+export { WASIElement } from "./elements/wasi";
+export { FileElement } from "./elements/file";
 export { RunnoProvider } from "./provider";
 export { headlessRunCode, headlessRunFS } from "./headless";
