+++ packages/runtime/src/provider.ts
@@ -1,5 +1,5 @@
-import { EditorElement } from "./editor";
-import { TerminalElement } from "./terminal";
+import { EditorElement } from "./elements/editor";
+import { TerminalElement } from "./elements/terminal";
 import {
   RunResult,
   WASIFS,
@@ -7,7 +7,7 @@ import {
   Syntax,
   RuntimeMethods,
 } from "@runno/host";
-import { ControlsElement } from "./controls";
+import { ControlsElement } from "./elements/controls";
 import { commandsForRuntime, getBinaryPathFromCommand } from "./commands";
 import { headlessPrepareFS, headlessRunCode, headlessRunFS } from "./headless";
 import { fetchWASIFS, makeRunnoError } from "./helpers";
